feat: simplify error boundaries
