+++ app/root.tsx
@@ -1,6 +1,5 @@
 import { json, LoaderFunctionArgs } from '@remix-run/node';
 import {
-    isRouteErrorResponse,
     Links,
     Meta,
     type MetaFunction,
@@ -8,18 +7,12 @@ import {
     Scripts,
     ScrollRestoration,
     useLoaderData,
-    useNavigate,
-    useNavigation,
-    useRouteError,
 } from '@remix-run/react';
 import { Tokens } from '@wix/sdk';
-import { useEffect } from 'react';
 import { CartOpenContextProvider } from '~/lib/cart-open-context';
 import { EcomApiContextProvider } from '~/lib/ecom';
 import { commitSession, initializeEcomSession } from '~/lib/ecom/session';
-import { getErrorMessage, routeLocationToUrl } from '~/lib/utils';
 import { RouteBreadcrumbs } from '~/src/components/breadcrumbs/use-breadcrumbs';
-import { ErrorPage } from '~/src/components/error-page/error-page';
 import { SiteWrapper } from '~/src/components/site-wrapper/site-wrapper';
 
 import '~/src/styles/reset.scss';
@@ -105,31 +98,4 @@ export default function App() {
     );
 }
 
-export function ErrorBoundary() {
-    const error = useRouteError();
-    const navigation = useNavigation();
-
-    useEffect(() => {
-        if (navigation.state === 'loading') {
-            const url = routeLocationToUrl(navigation.location, window.location.origin);
-            // force full page reload after navigating from error boundary
-            // to fix remix issue with style tags disappearing
-            window.location.assign(url);
-        }
-    }, [navigation]);
-
-    const navigate = useNavigate();
-
-    const isPageNotFoundError = isRouteErrorResponse(error) && error.status === 404;
-
-    return (
-        <ContentWrapper>
-            <ErrorPage
-                title={isPageNotFoundError ? 'Page Not Found' : 'Oops, something went wrong'}
-                message={isPageNotFoundError ? undefined : getErrorMessage(error)}
-                actionButtonText="Back to shopping"
-                onActionButtonClick={() => navigate('/products/all-products')}
-            />
-        </ContentWrapper>
-    );
-}
+export { ErrorBoundary } from '~/src/components/error-page/error-page';
