+++ app/routes/thank-you/route.tsx
@@ -1,9 +1,7 @@
 import type { LoaderFunctionArgs } from '@remix-run/node';
-import { useLoaderData, useRouteError } from '@remix-run/react';
+import { useLoaderData } from '@remix-run/react';
 import { initializeEcomApiForRequest } from '~/lib/ecom/session';
-import { getErrorMessage } from '~/lib/utils';
 import { CategoryLink } from '~/src/components/category-link/category-link';
-import { ErrorPage } from '~/src/components/error-page/error-page';
 import { OrderSummary } from '~/src/components/order-summary/order-summary';
 
 import styles from './route.module.scss';
@@ -39,7 +37,4 @@ export default function ThankYouPage() {
     );
 }
 
-export function ErrorBoundary() {
-    const error = useRouteError();
-    return <ErrorPage title="Error" message={getErrorMessage(error)} />;
-}
+export { ErrorBoundary } from '~/src/components/error-page/error-page';
