+++ lib/utils/common.ts
@@ -1,4 +1,4 @@
-import { isRouteErrorResponse, Location } from '@remix-run/react';
+import { isRouteErrorResponse } from '@remix-run/react';
 
 /**
  * It's important to add an appropriate role and a keyboard support
@@ -60,16 +60,6 @@ export function getErrorMessage(error: unknown): string {
     }
 
     return String(error);
-}
-
-/**
- * Converts Remix Location object into a standard URL object.
- */
-export function routeLocationToUrl(location: Location, origin: string): URL {
-    const url = new URL(location.pathname, origin);
-    url.search = location.search;
-    url.hash = location.hash;
-    return url;
 }
 
 /**
