+++ src/components/error-page/error-page.module.scss
@@ -1,6 +1,6 @@
 .root {
     padding: 48px 0;
-    min-height: 50vh;
+    min-height: 70vh;
     max-width: 40em;
     margin: 0 auto;
     display: flex;
